+++ packages/web/src/worker/sync/SharedSyncImplementation.ts
@@ -239,7 +239,7 @@ export class SharedSyncImplementation
    */
   async connect(options?: PowerSyncConnectionOptions) {
     this.lastConnectOptions = options;
-    return this.connectionManager.connect(CONNECTOR_PLACEHOLDER, options);
+    return this.connectionManager.connect(CONNECTOR_PLACEHOLDER, options ?? {});
   }
 
   async disconnect() {
@@ -318,7 +318,7 @@ export class SharedSyncImplementation
       this.dbAdapter = null;
 
       if (shouldReconnect) {
-        await this.connectionManager.connect(CONNECTOR_PLACEHOLDER, this.lastConnectOptions);
+        await this.connectionManager.connect(CONNECTOR_PLACEHOLDER, this.lastConnectOptions ?? {});
       }
     }
 
@@ -371,7 +371,7 @@ export class SharedSyncImplementation
     const syncParams = this.syncParams!;
     // Create a new StreamingSyncImplementation for each connect call. This is usually done is all SDKs.
     return new WebStreamingSyncImplementation({
-      adapter: new SqliteBucketStorage(this.dbAdapter!, new Mutex(), this.logger),
+      adapter: new SqliteBucketStorage(this.dbAdapter!, this.logger),
       remote: new WebRemote(
         {
           invalidateCredentials: async () => {