+++ packages/common/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @powersync/common
 
+## 1.34.0
+
+### Minor Changes
+
+- ab33799: Add experimental support for raw tables, giving you full control over the table structure to sync into.
+  While PowerSync manages tables as JSON views by default, raw tables have to be created by the application
+  developer.
+
+  For more information about raw tables, see [the documentation](https://docs.powersync.com/usage/use-case-examples/raw-tables).
+
+- 810c6ad: Propagate logger from PowerSyncDatabase to streaming sync and remote implementations, and tweak some log messages.
+
+### Patch Changes
+
+- a1aa18c: Fix sync stream delays during CRUD upload.
+- 9fb898d: [Internal] Removed shared mutex implementation of `readLock` and `writeLock`.
+
 ## 1.33.2
 
 ### Patch Changes